requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
Jinja2>=3.1.0
